myproject/requirements.txt
asgiref==3.8.1
Django==5.0.7
django-filter==24.2
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
greenlet==3.0.3
mysql-connector-python==9.0.0
mysqlclient==2.2.4
numpy==2.0.1
pandas==2.2.2
Python==3.10.12
PyJWT==2.8.0
python-dateutil==2.9.0.post0
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.31
sqlparse==0.5.1
typing_extensions==4.12.2
tzdata==2024.1
